refactor: VoteParticipationQueryService/VoteQueryService를 VoteService로 통합
