--- src/main/java/com/ject/vs/vote/port/VoteQueryService.java
@@ -1,17 +1,0 @@
-package com.ject.vs.vote.port;
-
-import com.ject.vs.vote.port.in.VoteQueryUseCase;
-import com.ject.vs.vote.port.in.dto.VoteStatus;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class VoteQueryService implements VoteQueryUseCase {
-
-    @Override
-    public List<Long> findAllVoteIdsByStatus(List<Long> voteIds, VoteStatus status) {
-        // TODO: Vote 도메인 연동 후 실제 status 기반 필터링으로 교체
-        return voteIds;
-    }
-}
