+++ src/test/java/com/ject/vs/vote/port/VoteServiceTest.java
@@ -2,6 +2,7 @@ package com.ject.vs.vote.port;
 
 import com.ject.vs.vote.domain.VoteParticipation;
 import com.ject.vs.vote.domain.VoteParticipationRepository;
+import com.ject.vs.vote.port.in.dto.VoteStatus;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -16,10 +17,10 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-class VoteParticipationQueryServiceTest {
+class VoteServiceTest {
 
     @InjectMocks
-    private VoteParticipationQueryService voteParticipationQueryService;
+    private VoteService voteService;
 
     @Mock
     private VoteParticipationRepository voteParticipationRepository;
@@ -30,12 +31,11 @@ class VoteParticipationQueryServiceTest {
         @Test
         void 유저가_참여한_voteId_목록을_반환한다() {
             // given
-            VoteParticipation p1 = VoteParticipation.of(10L, 1L);
-            VoteParticipation p2 = VoteParticipation.of(20L, 1L);
-            given(voteParticipationRepository.findAllByUserId(1L)).willReturn(List.of(p1, p2));
+            given(voteParticipationRepository.findAllByUserId(1L))
+                    .willReturn(List.of(VoteParticipation.of(10L, 1L), VoteParticipation.of(20L, 1L)));
 
             // when
-            List<Long> result = voteParticipationQueryService.findAllVoteIdsByUserId(1L);
+            List<Long> result = voteService.findAllVoteIdsByUserId(1L);
 
             // then
             assertThat(result).containsExactly(10L, 20L);
@@ -47,7 +47,7 @@ class VoteParticipationQueryServiceTest {
             given(voteParticipationRepository.findAllByUserId(1L)).willReturn(List.of());
 
             // when
-            List<Long> result = voteParticipationQueryService.findAllVoteIdsByUserId(1L);
+            List<Long> result = voteService.findAllVoteIdsByUserId(1L);
 
             // then
             assertThat(result).isEmpty();
@@ -63,7 +63,7 @@ class VoteParticipationQueryServiceTest {
             given(voteParticipationRepository.countByVoteId(1L)).willReturn(25L);
 
             // when
-            long result = voteParticipationQueryService.countParticipantsByVoteId(1L);
+            long result = voteService.countParticipantsByVoteId(1L);
 
             // then
             assertThat(result).isEqualTo(25L);
@@ -77,15 +77,43 @@ class VoteParticipationQueryServiceTest {
         @Test
         void 투표에_참여한_userId_목록을_반환한다() {
             // given
-            VoteParticipation p1 = VoteParticipation.of(1L, 100L);
-            VoteParticipation p2 = VoteParticipation.of(1L, 200L);
-            given(voteParticipationRepository.findAllByVoteId(1L)).willReturn(List.of(p1, p2));
+            given(voteParticipationRepository.findAllByVoteId(1L))
+                    .willReturn(List.of(VoteParticipation.of(1L, 100L), VoteParticipation.of(1L, 200L)));
 
             // when
-            List<Long> result = voteParticipationQueryService.findAllUserIdsByVoteId(1L);
+            List<Long> result = voteService.findAllUserIdsByVoteId(1L);
 
             // then
             assertThat(result).containsExactly(100L, 200L);
+        }
+    }
+
+    @Nested
+    class findAllVoteIdsByStatus {
+
+        @Test
+        void ONGOING_상태로_조회하면_전달된_voteId_목록을_그대로_반환한다() {
+            // given
+            List<Long> voteIds = List.of(1L, 2L, 3L);
+
+            // when
+            List<Long> result = voteService.findAllVoteIdsByStatus(voteIds, VoteStatus.ONGOING);
+
+            // then
+            // TODO: Vote 도메인 연동 후 실제 status 필터링 검증으로 교체
+            assertThat(result).isEqualTo(voteIds);
+        }
+
+        @Test
+        void 빈_목록을_전달하면_빈_목록을_반환한다() {
+            // given
+            List<Long> voteIds = List.of();
+
+            // when
+            List<Long> result = voteService.findAllVoteIdsByStatus(voteIds, VoteStatus.ENDED);
+
+            // then
+            assertThat(result).isEmpty();
         }
     }
 }
